requirements.txt
Flask
requests
python-dotenv
Pillow
Werkzeug
gunicorn
